refactor: move config to data folder
