+++ src/app/admin/sidebar.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { signOut } from "next-auth/react";
 import { cn } from "@/lib/utils";
-import { links } from "@/config";
+import { links } from "@/data/config";
 import { LogOutIcon } from "lucide-react";
 
 export const Sidebar = () => {
