+++ src/data/config.tsx
@@ -1,6 +1,6 @@
 import { BookIcon, UsersIcon } from "lucide-react";
 
-const links = [
+export const links = [
   {
     href: "/admin/books",
     label: "Шығармалар",
@@ -13,11 +13,9 @@ const links = [
   },
 ];
 
-const filters = {
+export const filters = {
   programs: ["JBBM", "NIS"],
   grades: ["5", "6", "7", "8", "9", "10", "11", "12"],
   languages: ["T1", "T2"],
   terms: ["1", "2", "3", "4"],
 };
-
-export { links, filters };
